% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap.R
\name{CI.FDF}
\alias{CI.FDF}
\title{CI.FDF}
\usage{
CI.FDF(fit, xarg, n.sample = NULL, n.rep = 100)
}
\arguments{
\item{fit}{Ouput fit from \code{FDF} function.}

\item{xarg}{The \code{CLRdata$x.fine} found in output from \code{PNSDdata_fd} function.}

\item{n.sample}{Original data sample size (used to limit the range of the bootstrap). 
No need if raw data is large enough.}

\item{n.rep}{Size of bootstrap samples.}
}
\value{
A list with: 
\itemize{
  \item \code{Bootstrap.sample} - an array of dimension length(xarg) x #factors x n.rep 
  containing the bootstrapped sample
  \item \code{x} - same as \code{xarg} from input
  \item \code{lower} - bootstrapped lower CI for all factors in fd format
  \item \code{upper} - bootstrapped upper CI for all factors in fd format
}
}
\description{
Computes the Bootstrapped confidence intervals for the PFF model
}
\examples{
library(fda)
data(fd.data2011)
data(CLRData2011)
m<- 2
factor_output1 <- FDF(data=fd.data2011,h=24*5,k=m,kmax = 10, 
                      kern_type = 'BT',plot=TRUE)
example.bootstrap <- CI.FDF(factor_output1,xarg=CLRData2011$x.fine,
                                   n.rep = 10)

plot(factor_output1$hat.F, lty = 1)
plot(example.bootstrap$lower, lty = 2, add = TRUE)
plot(example.bootstrap$upper, lty = 2, add = TRUE)

}
