% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIPFF.R, R/CIPFF2.R
\name{CI.PFF}
\alias{CI.PFF}
\title{CI.PFF}
\usage{
CI.PFF(factor_CIs, fit.PFF)

CI.PFF(factor_CIs, fit.PFF)
}
\arguments{
\item{factor_CIs}{output object from \code{CI.FDF}}

\item{fit.PFF}{output object from \code{PFF}

  A list with: 
\itemize{
  \item \code{LCIdf} - a data.frame with bootstrapped lower CI for all factors
  \item \code{UCIdf} - a data.frame with bootstrapped upper CI for all factors
}}
}
\description{
Convert the Bootstrapped confidence intervals for the FDF model into 
confidence intervals for the PFF model

Convert the Bootstrapped confidence intervals for the FDF model into 
confidence intervals for the PFF model
}
\examples{
library(fda)
data(fd.data2011)
data(CLRData2011)
m<- 2

factor_output1 <- FDF(data=fd.data2011,h=24*5,k=m,kmax = 10, 
                      kern_type = 'BT')
PFF.output<-PFF(factor_output1,dates,CLRData,fd.data)

set.seed(5643)
factor_CIs<-CI.FDF(factor_output1,xarg=CLRData2011$x.fine,n.rep = 10)
Sources_CIs<-CI.PFF(factor_CIs=factor_CIs,fit.PFF=PFF.output)

matplot(PFF.output$SourceProfiles, main='Sources',type='l',lty=1)
matplot(Sources_CIs$LCIdf[,-1], type='l', lty = 2, add = TRUE)        
matplot(Sources_CIs$UCIdf[,-1], type='l', lty = 2, add = TRUE)    
    
library(fda)
data(fd.data2011)
data(CLRData2011)
m<- 2

factor_output1 <- FDF(data=fd.data2011,h=24*5,k=m,kmax = 10, 
                      kern_type = 'BT')
PFF.output<-PFF(factor_output1,dates,CLRData,fd.data)

set.seed(5643)
factor_CIs<-CI.FDF(factor_output1,xarg=CLRData2011$x.fine,n.rep = 10)
Sources_CIs<-CI.PFF(factor_CIs=factor_CIs,fit.PFF=PFF.output)

matplot(PFF.output$SourceProfiles, main='Sources',type='l',lty=1)
matplot(Sources_CIs$LCIdf[,-1], type='l', lty = 2, add = TRUE)        
matplot(Sources_CIs$UCIdf[,-1], type='l', lty = 2, add = TRUE)    
    
}
