% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap.R
\name{CI_Bootstrap}
\alias{CI_Bootstrap}
\title{CI_Bootstrap}
\usage{
CI_Bootstrap(fit, xarg, n.sample = NULL, n.rep = 100)
}
\arguments{
\item{fit}{Ouput fit from \code{FDF} function.}

\item{xarg}{The \code{CLRdata$x.fine} found in output from \code{PNSDdata_fd} function.}

\item{n.sample}{Original data sample size (used to limit the range of the bootstrap). No need if raw data is large enough.}

\item{n.rep}{Size for of bootstrap samples.}
}
\value{
A list with: 
\itemize{
  \item matrix - the estimated time series
}
- \emph{matrix}: length(xarg)

x num_factors containing the sd evaluated at xarg
   x: a vector containing values where sd is evaluated
}
\description{
Compute the Bootstrapped Confidence Intervals
}
\examples{
 example.bootstrap <-
   CI.Bootstrap(fit = factor.hat,
                xarg,
                n.sample = 2000,
                n.rep = 100)

 plot(factor.hat$hat.F, lty = 1)
 plot(example.bootstrap$lower, lty = 2, add = TRUE)
 plot(example.bootstrap$upper, lty = 2, add = TRUE)

}
