% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{clr2density}
\alias{clr2density}
\title{clr2density: Invert the clr transformation}
\usage{
clr2density(z, z_step, clr)
}
\arguments{
\item{z}{Grid of points defining the abscissa}

\item{z_step}{Step for the grid of the abscissa}

\item{clr}{Grid evaluation of the clr transformed density}
}
\value{
Grid evaluation of the density
}
\description{
clr2density: Invert the clr transformation
}
\examples{
rnorm(1)
}
