% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{data.fd}
\alias{data.fd}
\title{data.fd}
\usage{
data.fd(
  argvals = NULL,
  data,
  basis,
  nbasis = 25,
  rangeval = c(0, 1),
  lambda = NULL
)
}
\arguments{
\item{argvals}{fill in}

\item{data}{fill in}

\item{basis}{fill in}

\item{nbasis}{fill in}

\item{rangeval}{fill in}

\item{lambda}{fill in}
}
\value{
fill in
}
\description{
Form a `fd` object from input data.
}
\examples{
rnorm(1)
}
