% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{factor_to_sources}
\alias{factor_to_sources}
\title{factor_to_sources}
\usage{
factor_to_sources(factor_hat, CLR_data, fd_data)
}
\arguments{
\item{factor_hat}{??}

\item{CLR_data}{??}

\item{fd_data}{??}
}
\value{
list
}
\description{
Function to convert factors back into sources
}
\examples{
rnorm(1)
}
