% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{FactorCIToSourcesCI}
\alias{FactorCIToSourcesCI}
\title{Convert Factor CIs to Source CIs}
\usage{
FactorCIToSourcesCI(factor_CIs, Res.FTS)
}
\arguments{
\item{factor_CIs}{fill in}

\item{Res.FTS}{fill in}
}
\value{
list
}
\description{
Convert Factor CIs to Source CIs
}
\examples{
rnorm(1)
}
