% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hatK.R
\name{hatK}
\alias{hatK}
\title{hatK}
\usage{
hatK(evalues)
}
\arguments{
\item{evalues}{Numeric vector of eigenvalues}
}
\value{
A list with: 
\itemize{
\item \code{hat_k}:  Index of the best k
\item \code{ratio_k}:  values of the ratio
}
}
\description{
Computes the auxiliary K, as the ratio of each eigenvalue to its predecessor;
returns the index and the ratio values.
}
\examples{
x <- rnorm(10)

}
