% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_Kernel.R
\name{Kernel}
\alias{Kernel}
\title{Kernel}
\usage{
Kernel(i, h, kerneltype)
}
\arguments{
\item{i}{fill in}

\item{h}{fill in}

\item{kerneltype}{Character input specifying type of kernel from list: `flat`, `Simple`, `Bartlett`, `flat_top` and `Parzen`}
}
\value{
Value of specified kernel for inputs `i` and `h`.
}
\description{
Given an input parameter `kerneltype`, returns the functional form of the kernel
as a function of inputs `i` and `h`: `x = i/h`.
}
\examples{
Kernel(i = 1, h = 2, kerneltype = "Bartlett")
}
