% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrc.R
\name{lrc}
\alias{lrc}
\title{lrc Long-Run Covariance}
\usage{
lrc(datafd, kern_type, h)
}
\arguments{
\item{datafd}{Input data, in functional data object.}

\item{kern_type}{Kernel type, switched internally to \code{c("BT", "PR", "SP", "FT", "flat")} which are Bartlett, Parzen, Simple, flat_top or flat kernels.}

\item{h}{The bandwidth parameter, strictly non-zero.}
}
\value{
list
}
\description{
Estimates the long-run covariance operator, as discussed originally in
Rice, G., and Shang, H.L. (2017), "A Plug-in Bandwidth Selection Procedure for Long-run 
Covariance Estimation with Stationary Functional Time Series", 
Journal of Time Series Analysis, 38, 591–609.
}
