% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrc.R
\name{lrc_paper}
\alias{lrc_paper}
\title{lrc_paper Long-Run Covariance, Variant 2}
\usage{
lrc_paper(datafd, kern_type, h, p)
}
\arguments{
\item{datafd}{Input data, in FD format.}

\item{kern_type}{Kernel type, switched internally to \code{c("BT", "PR", "SP", "FT", "flat")} which are Bartlett, Parzen, Simple, flat_top or flat kernels.}

\item{h}{The bandwidth parameter, strictly non-zero.}

\item{p}{fill in}
}
\value{
list
}
\description{
Estimates the long-run covariance, as discussed originally in 
[Israel's paper]
}
\examples{
rnorm(1)
}
