% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrc.R
\name{lrc_paper_NoInv}
\alias{lrc_paper_NoInv}
\title{Estimation of the Lambda operator - Long-run covariance}
\usage{
lrc_paper_NoInv(datafd, kern_type, h, replicates = NULL)
}
\arguments{
\item{datafd}{Input data, in FD format.}

\item{kern_type}{Kernel type, switched internally to \code{c("BT", "PR", "SP", "FT", "flat")} which are Bartlett, Parzen, Simple, flat_top or flat kernels.}

\item{h}{The bandwidth parameter, strictly non-zero.}

\item{replicates}{Number of replicates (default NULL) for the ...}
}
\value{
A list with:
\itemize{
  \item \code{eigenf} - eigenfunctions of the operator
  \item \code{eigenval} - eigenvalues of the operator
  \item \code{lrc.k} - kernel of the lambda operator
}
}
\description{
Estimates the long-run covariance operator lambda described in the main paper.
}
\details{
This uses the method discussed originally in
Rice, G., and Shang, H.L. (2017), "A Plug-in Bandwidth Selection Procedure for Long-run 
Covariance Estimation with Stationary Functional Time Series", 
Journal of Time Series Analysis, 38, 591–609.
}
