% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFF.R
\name{PFF}
\alias{PFF}
\title{PFF}
\usage{
PFF(factor_hat, dates, CLR_data, fd_data, flip = TRUE)
}
\arguments{
\item{factor_hat}{Output from FDF.R}

\item{dates}{Data dates}

\item{CLR_data}{Output from PNSDdata.fd}

\item{fd_data}{Output from PNSDdata.fd}

\item{flip}{tuning parameter to decide if factors should be flipped or not based on 
dispersion around main mode.}
}
\value{
A list with: 
\itemize{
  \item \code{SourceProfiles} - a data.frame with the estimated densities
  \item \code{SourceContributions} - a data.frame with dates and time series contributions based on interpretable projection
  \item \code{SourceContributions0} - a data.frame with dates and time series contributions based on model fit
  \item \code{RowSumSourceContribution} - normalization constant in \code{SourceContributions}
  \item \code{log.x, x.step} - values used during density normalization in log range 
  \item \code{signs} - (not for users)
  \item \code{sig2} - (not for users)
}
}
\description{
After fitting a factor model (output from FDF) in Hilbert space, 
it transforms factors and contributions to original density space
}
\examples{
library(fda)
data(fd.data2011);data(CLRData2011)
m<- 5
factor_output1 <- FDF(data=fd.data2011,h=24*5,k=m,kmax = 10, 
                      kern_type = 'BT')
PFF.output<-PFF(factor_output1,dates,CLRData,fd.data)
                      
matplot(PFF.output$SourceProfiles, main='Sources')
matplot(PFF.output$SourceContributions[,-1], type='l', 
         main=expression(paste('Time series ',beta[t])),
         xlab= 'time') 
}
