% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PNSDdatafd.R
\name{PNSDdata.fd}
\alias{PNSDdata.fd}
\title{PNSDdata.fd}
\usage{
PNSDdata.fd(Dens, x, n.basis = 20, k = 4, der = 2, alpha = 0.999, ch = 1)
}
\arguments{
\item{Dens}{A matrix, `m x n`, where `m` is the number of PNSD.}

\item{x}{A vector with particle number sizes grid.}

\item{k}{Order of spline, degree = k-1}

\item{der}{Derivation (see smoothing_spline0)}

\item{alpha}{Smoothing parameter}

\item{ch}{See smoothing_spline0}

\item{nbasis}{Number of basis functions to be used for the functional data.}

\item{lambda}{Non-negative real number, smoothing parameter to be applied to the estimated functional parameter. If NULL, estimated using generalized CV.}
}
\value{
A list with: 
\itemize{
  \item \code{CLRData} - parameters concerning preprocess and transformation to clr data
  \item \code{fd.data} - functional representation of densities in Hilbert space
}
}
\description{
Preprocess raw PNSD to functional representation for densities
}
\examples{
data(Dens2011)
data(ps2011)
Test<-PNSDdata.fd(Dens2011[11:15,],ps2011)
yy<-eval.fd(Test$CLRData$x.fine,Test$fd.data)
for(i in 1:ncol(yy)){
yy[,i]<-clr2density(Test$CLRData$x.fine,z_step = Test$CLRData$x.step,yy[,i])
}

par(mfrow=c(2,1))
matplot(ps2011,t(Test$CLRData$densities),type='p',pch=20,ylim=c(0,1),main="Raw Densities",ylab='',xlab='Particle size')
matplot(exp(Test$CLRData$x.fine),yy,type='l',ylim=c(0,1),main="Functional Densities",ylab='',xlab='Particle size')


}
