% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{run_flip}
\alias{run_flip}
\title{run_flip}
\usage{
run_flip(SourceProfiles, SourceContributions, log.x)
}
\arguments{
\item{SourceProfiles}{fill in}

\item{SourceContributions}{fill in}

\item{log.x}{fill in}
}
\value{
list
}
\description{
run_flip
}
