% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing_spline0.R
\name{smoothing_spline0}
\alias{smoothing_spline0}
\title{Modified Smoothing Spline}
\usage{
smoothing_spline0(knots, tp, f, w, k, der, alpha, ch = FALSE)
}
\arguments{
\item{knots}{Input knot list for spline basis creation.}

\item{tp}{Points of approximation.}

\item{f}{Values at tp.}

\item{w}{Weight coefficients.}

\item{k}{Order of the B-spline basis, one higher than their degree.}

\item{der}{Derivation?}

\item{alpha}{Smoothing parameter.}

\item{ch}{if FALSE, then functional with (1-alpha) and alpha; if TRUE, then functional with alpha alone.}
}
\value{
A list of three objects: `list(J, z, spline0))`
}
\description{
Method described in Hron, Menafoglio, Templ, Hruzova and Filzmoser (2016),
`Simplicial principal component analysis for density functions in Bayes spaces`,
Computational Statistics & Data Analysis, Volume 94, Pages 330-350, ISSN 0167-9473.'
Extension of the original grid by adding (k-1) knots.
}
