% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{sumF}
\alias{sumF}
\title{sumF}
\usage{
sumF(Fmat, s)
}
\arguments{
\item{Fmat}{}

\item{s}{Observation times; first difference is the base for the sum.}
}
\value{
The normalized vector of products.
}
\description{
Act across the columns of `Fmat`, and compute the `sumfg` on
the columns pairwise, starting from the first two. Continue
doing so until all columns are multiplied. Normalizes each product.
}
\examples{
rnorm(1)
}
