% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{sumfg}
\alias{sumfg}
\title{sumfg}
\usage{
sumfg(f, g, s)
}
\arguments{
\item{f}{Input 1}

\item{g}{Input 2}

\item{s}{Observation times; first difference is the base for the sum.}
}
\value{
A list of two objects: the `sum` and the `norm` (constant)
}
\description{
Utility function to compute the midpoint-rule Riemann sum of `f * g`
with possibly irregular bases (from the first-difference of `s`) and 
then return the product of `f * g`, normalized by that sum.
}
\examples{
rnorm(1)
}
