% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{trapzc}
\alias{trapzc}
\title{traczc - Trapezoidal Integration}
\usage{
trapzc(step, y)
}
\arguments{
\item{step}{Step of the grid.}

\item{y}{Vector of points to be integrated.}
}
\value{
Numerical integration via trapezoidal formula.
}
\description{
Perform a trapezoidal Riemann sum to the input `y`, with step `step`
}
