% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{TS.contribution}
\alias{TS.contribution}
\title{TS.contribution}
\usage{
TS.contribution(
  SourceProfiles2,
  SourceContributions2,
  fd.data,
  CLR_data,
  log.x
)
}
\arguments{
\item{SourceProfiles2}{fill in}

\item{SourceContributions2}{fill in}

\item{fd.data}{fill in}

\item{CLR_data}{fill in}

\item{log.x}{fill in}
}
\value{
fill in
}
\description{
Blah blah blah
}
\examples{
rnorm(1)
}
