% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_spline_basis.R
\name{z_spline_basis}
\alias{z_spline_basis}
\title{z_spline_basis: Computes a basis set of Z-splines}
\usage{
z_spline_basis(knots, k)
}
\arguments{
\item{knots}{Input knot list for spline basis creation.}

\item{k}{Order of the B-spline basis, one higher than their degree.}
}
\value{
list(C0 = C0, M0 = M, K = L, D = Dmat)
  Four matrix objects.
}
\description{
Method described in Hron, Menafoglio, Templ, Hruzova and Filzmoser (2016),
`Simplicial principal component analysis for density functions in Bayes spaces`,
Computational Statistics & Data Analysis, Volume 94, Pages 330-350, ISSN 0167-9473.
}
\examples{
  knots <- seq(from = 1, to = 20, length.out = 100)
  k <- 4
  res <- z_spline_basis(knots = knots, k = k)
}
