% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{factor_to_sources}
\alias{factor_to_sources}
\title{factor_to_sources}
\usage{
factor_to_sources(factor_hat, dates, CLR_data, fd_data, flip = TRUE)
}
\arguments{
\item{factor_hat}{Output from FDF.R}

\item{dates}{Data dates}

\item{CLR_data}{Output from PNSDdata_fd}

\item{fd_data}{Output from PNSDdata_fd}
}
\value{
list()
  SourceProfiles: densities 
  SourceContributions: time series contributions 
  and other metadata
}
\description{
After fitting a factor model (output from FDF) it transforms factors and 
contributions to original density space
}
\examples{
see document 
}
