% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrc.R
\name{lrc}
\alias{lrc}
\title{lrc Long-Run Covariance}
\usage{
lrc(datafd, kern_type, h)
}
\arguments{
\item{datafd}{Input data, in FD format.}

\item{kern_type}{Kernel type, switched internally to \code{c("BT", "PR", "SP", "FT", "flat")} which are Bartlett, Parzen, Simple, flat_top or flat kernels.}

\item{h}{The bandwidth parameter, strictly non-zero.}
}
\value{
list
}
\description{
Estimates the long-run covariance operator, as discussed originally in
Aue, A., Rice, G. and Sönmez, O. "Detecting and dating structural breaks in functional data without dimension reduction." Journal of the Royal Statistical Society Series B: Statistical Methodology 80.3 (2018): 509-529.
}
\examples{
rnorm(1)
}
