% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrc.R
\name{lrc_paper_NoInv}
\alias{lrc_paper_NoInv}
\title{lrc_paper_NoInv Long-Run Covariance, Variant 2, No Inverse}
\usage{
lrc_paper_NoInv(datafd, kern_type, h, replicates = NULL)
}
\arguments{
\item{datafd}{Input data, in FD format.}

\item{kern_type}{Kernel type, switched internally to \code{c("BT", "PR", "SP", "FT", "flat")} which are Bartlett, Parzen, Simple, flat_top or flat kernels.}

\item{h}{The bandwidth parameter, strictly non-zero.}

\item{replicates}{Number of replicates (default NULL) for the ...}
}
\value{
list
}
\description{
Estimates the long-run covariance, as discussed originally in 
[Israel's paper]. This is the variant without the inverse.
}
\examples{
rnorm(1)
}
