% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{prodcf}
\alias{prodcf}
\title{prodcf}
\usage{
prodcf(c, f, s)
}
\arguments{
\item{c}{Power}

\item{f}{Base}

\item{s}{Observation times; first difference is the base for the sum.}
}
\value{
A list of two objects: the `sum` and the `norm` (constant)
}
\description{
Utility function to compute the midpoint-rule Riemann sum of `f^c`
with possibly irregular bases (from the first-difference of `s`) and 
then return the product of `f^c`, normalized by that sum.
#'
}
\examples{
ss<-seq(-4,5,length.out=100) 
f<-dnorm(ss,mean=0,sd=1) alp<-2
alp.times.f<-prodcf(alp,f=f,s=ss)$cf
plot(ss,f,type="l",lwd=2,col="orange",ylim=c(0,max(f,alp.times.f)))
lines(ss,alp.times.f,lwd=2,col="green")
}
