% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{sumfg}
\alias{sumfg}
\title{sumfg}
\usage{
sumfg(f, g, s)
}
\arguments{
\item{f}{Input 1}

\item{g}{Input 2}

\item{s}{Observation times; first difference is the base for the sum.}
}
\value{
A list of two objects: the `sum` and the `norm` (constant)
}
\description{
Utility function to compute the midpoint-rule Riemann sum of `f * g`
with possibly irregular bases (from the first-difference of `s`) and 
then return the product of `f * g`, normalized by that sum.
}
\examples{
ss<-seq(-4,5,length.out=100)
f<-dnorm(ss,mean=0,sd=1)
g<-dnorm(ss,mean=1,sd=sqrt(2))
fpg<-sumfg(f,g,ss)$sum
plot(ss,f,type="l",lwd=2,col="orange",ylim=c(0,max(f,g,fpg)),ylab="",main="f plus g")
lines(ss,g,lwd=2,col="blue")
lines(ss,fpg,lwd=2,col="green")
legend('topright',legen=c("f","g"),col=c("orange","blue"),lty=1)
}
